Recruiter_UI/requirements.txt
playwright>=1.50.0
pytest>=8.0.0
pytest-html>=4.1.0
pytest-metadata>=3.0.0
pytest-json-report>=1.5.0
python-dotenv>=1.0.0
